% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_streamline.R
\name{asStreamline}
\alias{asStreamline}
\title{Create a streamline object}
\usage{
asStreamline(line, seedIndex = NULL, spaceDims = NULL, voxelDims = NULL,
  image = NULL, coordUnit = c("vox", "mm"))
}
\arguments{
\item{line}{A numeric matrix with three columns, giving the (x,y,z)
coordinates of each point on the line, one per row.}

\item{seedIndex}{An integer giving the index (row of \code{line}) of the seed
point from which the streamline was generated.}

\item{spaceDims}{An integer vector of length 3 giving the dimensions of
the image space that the streamline was generated within. If \code{NULL},
this and \code{voxelDims} are obtained from \code{image} instead.}

\item{voxelDims}{A numeric vector of length 3 giving the voxel dimensions
of the image space that the streamline was generated within. If \code{NULL},
this and \code{spaceDims} are obtained from \code{image} instead.}

\item{image}{An image, or path to one, from which \code{spaceDims} and
\code{voxelDims} can be obtained if they are not given explicitly.}

\item{coordUnit}{A string giving the unit that the coordinates in \code{line}
are given in, either \code{"vox"} (voxels) or \code{"mm"} (millimetres).}
}
\value{
A \link{Streamline} object.
}
\description{
Create a \link{Streamline} object from a matrix of points, checking and
completing metadata about the image space as required.
}
\author{
Jon Clayden
}
