% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_streamline.R
\name{attachStreamlines}
\alias{attachStreamlines}
\title{Attach streamlines held in memory to a source}
\usage{
attachStreamlines(streamlines)
}
\arguments{
\item{streamlines}{A \link{Streamline} object, or a list of them.}
}
\value{
A \link{StreamlineSource} object encapsulating the given streamlines.
}
\description{
Wrap one or more \link{Streamline} objects already held in memory in a
\link{StreamlineSource}, so that they can be filtered, selected and processed
using the same interface as streamlines read from file or generated by
tracking.
}
\author{
Jon Clayden
}
