% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_model.R
\name{bedpostDiffusionModel}
\alias{bedpostDiffusionModel}
\title{Create a ball-and-sticks diffusion model}
\usage{
bedpostDiffusionModel(bedpostDir, avfThreshold = 0.05)
}
\arguments{
\item{bedpostDir}{A character string giving the path to a BEDPOST(X)
output directory.}

\item{avfThreshold}{A numeric value giving the anisotropic volume fraction
below which a fibre compartment will be ignored during tractography.}
}
\value{
A \link{DiffusionModel} object of type \code{"bedpost"}.
}
\description{
Create a \link{DiffusionModel} representing a ball-and-sticks (BEDPOST) fit,
based on the contents of a BEDPOST(X) output directory.
}
\references{
Please cite the following reference when using TractoR in your
work:

J.D. Clayden, S. Muñoz Maniega, A.J. Storkey, M.D. King, M.E. Bastin & C.A.
Clark (2011). TractoR: Magnetic resonance imaging and tractography with R.
Journal of Statistical Software 44(8):1-18. \doi{10.18637/jss.v044.i08}.
}
\author{
Jon Clayden
}
