% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_tracker.R
\name{createTracker}
\alias{createTracker}
\title{Create a streamline tracker}
\usage{
createTracker(model, mask, curvatureThreshold = 0.2, loopcheck = TRUE,
  maxSteps = 2000, stepLength = 0.5, oneWay = FALSE)
}
\arguments{
\item{model}{A \link{DiffusionModel} object, as created by \code{\link[=dtiDiffusionModel]{dtiDiffusionModel()}}
or \code{\link[=bedpostDiffusionModel]{bedpostDiffusionModel()}}.}

\item{mask}{An optional mask image (or path to one), restricting the
region within which tracking is permitted.}

\item{curvatureThreshold}{A numeric value giving the minimum permitted dot
product between successive tracking steps, controlling how sharply a
streamline can turn.}

\item{loopcheck}{Logical value: should streamlines be terminated when they
start to loop back on themselves?}

\item{maxSteps}{An integer giving the maximum number of steps to take in
each direction from a seed point.}

\item{stepLength}{A numeric value giving the distance, in mm, moved at
each tracking step.}

\item{oneWay}{Logical value: if \code{TRUE}, tracking will proceed in only one
direction from each seed point, rather than the usual two.}
}
\value{
A \link{Tracker} object.
}
\description{
Create a \link{Tracker} object, which encapsulates a fitted diffusion model,
seeding mask and set of tracking parameters, and is used to generate
streamlines from seed points.
}
\references{
Please cite the following reference when using TractoR in your
work:

J.D. Clayden, S. Muñoz Maniega, A.J. Storkey, M.D. King, M.E. Bastin & C.A.
Clark (2011). TractoR: Magnetic resonance imaging and tractography with R.
Journal of Statistical Software 44(8):1-18. \doi{10.18637/jss.v044.i08}.
}
\author{
Jon Clayden
}
