% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_model.R
\docType{class}
\name{DiffusionModel-class}
\alias{DiffusionModel-class}
\alias{DiffusionModel}
\title{The DiffusionModel class}
\description{
This class represents a fitted diffusion model, encapsulating a pointer to
an underlying C++ object which is used by a \link{Tracker} to perform
streamline tractography. Objects of this class are created by
\code{\link[=dtiDiffusionModel]{dtiDiffusionModel()}} or \code{\link[=bedpostDiffusionModel]{bedpostDiffusionModel()}}, rather than directly.
}
\section{Fields}{

\describe{
\item{\code{pointer}}{An external pointer to the underlying C++ \code{DiffusionModel}
object.}

\item{\code{type}}{A string naming the type of model, currently either \code{"dti"} or
\code{"bedpost"}.}
}}


