% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_model.R
\name{dtiDiffusionModel}
\alias{dtiDiffusionModel}
\title{Create a diffusion tensor model}
\usage{
dtiDiffusionModel(directionsPath)
}
\arguments{
\item{directionsPath}{A character vector giving the path to the principal
diffusion direction image or file stem.}
}
\value{
A \link{DiffusionModel} object of type \code{"dti"}.
}
\description{
Create a \link{DiffusionModel} representing a diffusion tensor fit, based on an
image of principal diffusion directions.
}
\references{
Please cite the following reference when using TractoR in your
work:

J.D. Clayden, S. Muñoz Maniega, A.J. Storkey, M.D. King, M.E. Bastin & C.A.
Clark (2011). TractoR: Magnetic resonance imaging and tractography with R.
Journal of Statistical Software 44(8):1-18. \doi{10.18637/jss.v044.i08}.
}
\author{
Jon Clayden
}
