% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_streamline.R
\name{generateStreamlines}
\alias{generateStreamlines}
\title{Generate streamlines from seed points}
\usage{
generateStreamlines(tracker, seeds, countPerSeed, rightwardsVector = NULL,
  jitter = TRUE)
}
\arguments{
\item{tracker}{A \link{Tracker} object.}

\item{seeds}{A numeric matrix with three columns, giving the (x,y,z)
voxel coordinates of the seed points to track from, one per row.}

\item{countPerSeed}{An integer giving the number of streamlines to
generate from each seed point.}

\item{rightwardsVector}{An optional numeric vector of length 3, used to
disambiguate left-right symmetric seed neighbourhoods.}

\item{jitter}{Logical value: should seed points be randomly perturbed
within their voxel before tracking?}
}
\value{
A \link{StreamlineSource} object encapsulating the generated
streamlines.
}
\description{
Perform streamline tractography from a set of seed points, using a
previously created \link{Tracker}.
}
\references{
Please cite the following reference when using TractoR in your
work:

J.D. Clayden, S. Muñoz Maniega, A.J. Storkey, M.D. King, M.E. Bastin & C.A.
Clark (2011). TractoR: Magnetic resonance imaging and tractography with R.
Journal of Statistical Software 44(8):1-18. \doi{10.18637/jss.v044.i08}.
}
\author{
Jon Clayden
}
