% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_model.R
\name{getBedpostNumberOfFibres}
\alias{getBedpostNumberOfFibres}
\title{Count the number of fibres in a BEDPOST directory}
\usage{
getBedpostNumberOfFibres(bedpostDir)
}
\arguments{
\item{bedpostDir}{A character string giving the path to a BEDPOST(X)
output directory.}
}
\value{
An integer giving the number of fibres represented in the
directory.
}
\description{
Determine the number of fibre compartments modelled within a BEDPOST(X)
output directory, by counting the number of \verb{mean_f<n>samples} images
present.
}
\author{
Jon Clayden
}
