% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_model.R
\name{nilModel}
\alias{nilModel}
\title{Test for or create a nil diffusion model}
\usage{
nilModel(object = NULL)
}
\arguments{
\item{object}{An object, or \code{NULL}.}
}
\value{
If \code{object} is \code{NULL}, the nil model singleton is returned.
Otherwise, a logical value indicating whether \code{object} is identical to
the nil model.
}
\description{
A "nil" \link{DiffusionModel} is used to indicate the absence of a valid,
fitted diffusion model, for example when creating a \link{Tracker} that will
not itself be used to generate streamlines (as when reading them from
file instead).
}
\author{
Jon Clayden
}
