% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_model.R
\name{nilPointer}
\alias{nilPointer}
\title{Test for or create a nil pointer}
\usage{
nilPointer(object = NULL)
}
\arguments{
\item{object}{An object, or \code{NULL}.}
}
\value{
If \code{object} is \code{NULL}, the nil pointer singleton is returned.
Otherwise, a logical value indicating whether \code{object} is identical to
the nil pointer.
}
\description{
A "nil" external pointer is used throughout this package to indicate the
absence of a valid underlying C++ object, in contexts where an ordinary
\code{NULL} is not appropriate.
}
\author{
Jon Clayden
}
