% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_streamline.R
\name{readStreamlines}
\alias{readStreamlines}
\title{Read streamlines from file}
\usage{
readStreamlines(fileName, readLabels = TRUE)
}
\arguments{
\item{fileName}{A character string giving the path to the streamline
file, with or without an appropriate extension.}

\item{readLabels}{Logical value: should streamline labels, if present,
be read in along with the streamlines themselves?}
}
\value{
A \link{StreamlineSource} object encapsulating the streamlines in the
file.
}
\description{
Open a file of previously generated streamlines, in TrackVis (.trk) or
MRtrix (.tck) format, for subsequent processing.
}
\references{
Please cite the following reference when using TractoR in your
work:

J.D. Clayden, S. Muñoz Maniega, A.J. Storkey, M.D. King, M.E. Bastin & C.A.
Clark (2011). TractoR: Magnetic resonance imaging and tractography with R.
Journal of Statistical Software 44(8):1-18. \doi{10.18637/jss.v044.i08}.
}
\author{
Jon Clayden
}
