% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_tracker.R
\docType{class}
\name{Tracker-class}
\alias{Tracker-class}
\alias{Tracker}
\title{The Tracker class}
\description{
This class represents a streamline tractography engine, associated with a
particular \link{DiffusionModel}, seeding mask and set of tracking parameters.
It is used to generate streamlines from seed points, via
\code{\link[=generateStreamlines]{generateStreamlines()}}.
}
\section{Fields}{

\describe{
\item{\code{model}}{The \link{DiffusionModel} object used to guide tracking.}

\item{\code{pointer}}{An external pointer to the underlying C++ \code{Tracker} object.}
}}

\section{Methods}{

\describe{
\item{\code{setTargets(image, indices = NULL, labels = NULL, terminate = FALSE)}}{Set or clear target regions used to constrain or terminate streamlines}
}}

